리팩토링: 피연산자와 계산 결과값을 실수(double형)로 받도록 수정
- 기존에는 피연산자와 계산 결과값을 정수(int형)로만 받을 수 있었는데
실수 계산도 가능하도록 double형으로 확장

+++ src/main/java/calculator/Calculator.java
@@ -2,11 +2,11 @@ package calculator;
 
 public class Calculator {
 
-    public static int calculate(Expression expression) {
+    public static double calculate(Expression expression) {
         Operand firstOperand = new Operand(0+"");
         Operator operator = new Operator("+");
         Operand secondOperand = expression.getOperand(1);
-        int result = 0;
+        double result = 0;
 
         for (int i = 1; operator != null && secondOperand != null; i++) {
             result = calculateBinomial(firstOperand, operator, secondOperand);
@@ -19,14 +19,14 @@ public class Calculator {
         return result;
     }
 
-    private static int calculateBinomial(Operand firstOperand,
+    private static double calculateBinomial(Operand firstOperand,
             Operator operator, Operand secondOperand) {
         return calculateBinomialByRawData(firstOperand.getValue(),
             operator, secondOperand.getValue());
     }
 
-    private static int calculateBinomialByRawData(int firstOperand,
-            Operator operator, int secondOperand) {
+    private static double calculateBinomialByRawData(double firstOperand,
+            Operator operator, double secondOperand) {
 
         if (operator.isAdd()) {
             return firstOperand + secondOperand;
